feat: move to cjs builds to ensure compatibility with create-react-app
In order to maintain full compatibility with Create React App and its chosen ecosystem (Jest, et. al.) we will publish the UI library as CJS modules compiled to ES5, but without polyfills for ES6/7 features.

These polyfills are included by default with CRA, but for other environments the necessary polyfills needs to be included by the host app depending on the target environment. The polyfills listed match the target environments defined in `.browserslistrc`.