feat: version 2 of ui-core (#183)
feat: version 2.0.0

This release does not contain any new components over v1, but it does manage to change almost everything.

Features include:

- simpler components
- better composition
- better isolation from CSS leakage
- consistent component APIs
- better documentation
- better development tools when working on components

BREAKING CHANGE: The component API has been revamped to be simpler and more consistent. See the documentation for a full list.

BREAKING CHANGE: This changes the `list` prop passed to DropdownButton and SplitButton to `component`, which enables any component to be used as a menu.

BREAKING CHANGE: Event handlers align with the JS API of being called with one argument, the Event.