feat(buttonstrip): add button strip component (#227)
* fix(modal): spaces buttons

* chore(modal): remove rogue classNames from stories

* feat(button-strip): introduce ButtonStrip and use if for SplitButton

* fix(modal actions): undo changes to modal action in favor of ButtonStrip

* fix(modal): align Modal.Actions Buttons with ButtonStrip

* fix(button-strip): restore border-width to fix focus styles

* refactor: button focus styles
I am now using the ::after pseudo element for the focus outline. To me this makes sense for various reasons:
- This decouples the focus outline from the buttons border styles, so we can show an even outline around anything
- This solves rendering issues that can occur when trying to "fake" a thick border with a thin border + a box-shadow (especially on high-res screens)
- We wanted the outline to "stick out", this can be problematic because adjacent elements could overlap the "sticky-out-bit". In the previous box-shadow implementation this couldn't be prevented, but with this pseudo element it has been solved by adding a z-index.

* refactor(button-strip): use React.Children.count in prop validation

* docs(button-strip): documents usage of the ButtonStrip component

* fix: button styles
There was a problem in FireFox when buttons would get focus and active state simultaneously in the ButtonStrip. The problem was a result of CSS selector specificity: the `button.jsx-3548810694:active.jsx-3548810694:focus` selector is more specific than `div.compact.jsx-1219281741 > button:first-child`. And because the border style in the `button.jsx:active:focus` was defined as `border: 1px solid someColor;`, this would override the ButtonStrip styles and apply 1px on the right. To prevent this from happening, I am now only defining the border width in the base buttons styles, and only setting border-color in the rest of the styles.

* docs(button-strip): remove props-table and add info

* refactor(button-strip): use `start`, `middle` and `end` for alignment

* fix(button-strip): update modal stories to new API