refactor(select): update component to new design
To match the design of the Select in the Design System, the Select
component based on the native HTML select has to be deprecated and a
custom select component must be used.

This also adds components for multi-select and filterable select menus.

BREAKING CHANGE: Select deprecated in favor of the SingleSelect, refer
to the docs for the API.

BREAKING CHANGE: SelectField deprecated in favor of SingleSelectField,
see the docs for the API.

BREAKING CHANGE: The new Select components do not accept a standard HTML `option`
element, instead they require use of SingleSelectOption, MultiSelectOption,
or a component with a matching API.

BREAKING CHANGE: The Select's `onChange` no longer returns a standard `Event.target`,
instead returns the selected Object with properties: `label` and
`string`, or an array of those for the MultiSelects

BREAKING CHANGE: The Selects all require a selected prop, instead of a value.
Where selected is an object with a label and a value string, or an array of
those for the MultiSelect.

BREAKING CHANGE: Since the Selects do not use a native input element, the name
prop has been dropped from the Select components.