refactor(fileinput): align 'on' function handlers
Also fixes a bug where the current list is always returned for the
onChange function, but the current list is always empty because it never
returns the added item.

BREAKING CHANGE: `onChange` is called with two parameters:
First parameter an object with properties for `checked`, `name`, and
`value`.
The second is the React `event` object.