docs(api): add workaround for jsdoc bug
JSDoc has a bug where if a named import is the last statement before the
doc comment, then it fails to pick up the doc comment.

Any valid statement between the named import and the doc comment works
around the problem, such as `;` or `''`. Our code style rewrites a `''`
statement as `;('')`, and that is the shortest non-functional workaround
I have found to work around the problem.

Refs https://github.com/jsdoc/jsdoc/issues/1718
Refs https://github.com/jsdoc/jsdoc/issues/1706